refact: move Flow.await to :platform:ui
