+++ platform/ui/src/main/java/com/jeanbarrossilva/orca/platform/ui/core/Flow.extensions.kt
@@ -3,10 +3,25 @@ package com.jeanbarrossilva.orca.platform.ui.core
 import com.jeanbarrossilva.orca.platform.ui.core.replacement.emptyReplacementList
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.merge
 import kotlinx.coroutines.flow.runningFold
+
+/**
+ * Suspends until the next value is emitted to this [Flow] and then returns it.
+ *
+ * @param T Value emitted to this [Flow].
+ */
+suspend fun <T> Flow<T>.await(): T {
+  return if (this is StateFlow<T>) {
+    value.let { previous -> first { current -> current != previous } }
+  } else {
+    first()
+  }
+}
 
 /**
  * Maps each element of the emitted [Collection]s to the resulting [Flow] of [transform], merging
