+++ platform/ui/src/test/java/com/jeanbarrossilva/orca/platform/ui/core/FlowExtensionsTests.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.orca.feature.feed.viewmodel
+package com.jeanbarrossilva.orca.platform.ui.core
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
